Refactor TsvUtils
